requirements1.txt
alabaster==0.7.7
amqp==1.4.8
anyjson==0.3.3
appnope==0.1.0
arrow==0.7.0
Babel==2.2.0
billiard==3.3.0.22
bson==0.4.1
celery==3.1.19
cffi==1.3.1
click==6.4
configobj==5.0.6
coreapi==1.32.0
cryptography==1.1.2
decorator==4.0.6
Django==1.9.5
django-celery==3.1.17
django-jsonfield==0.9.15
django-redis-sessions==0.5.0
django-rest-swagger==0.3.8
django-suit==0.2.18
djangorestframework==3.3.3
docutils==0.12
enum34==1.1.1
gnureadline==6.3.3
idna==2.0
imagesize==0.7.1
ipaddress==1.0.15
ipython==4.0.1
ipython-genutils==0.1.0
iso8601==0.1.11
itypes==1.1.0
Jinja2==2.8
jpush==3.2.1
jsonfield==1.0.3
kombu==3.0.32
leancloud-sdk==1.3.10
MarkupSafe==0.23
mongoengine==0.10.6
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
openapi-codec==1.0.0
path.py==8.1.2
pexpect==4.0.1
pgcli==0.20.1
pgspecial==1.3.0
phonenumbers==7.2.3
pickleshare==0.5
prompt-toolkit==0.46
psycopg2==2.6.1
ptyprocess==0.5
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1
pymongo==3.2
PyMySQL==0.6.7
pyOpenSSL==0.15.1
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
qiniu==7.0.6
redis==2.10.5
requests==2.8.1
setproctitle==1.1.9
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.5
sphinx-rtd-theme==0.1.9
sqlparse==0.1.16
tabulate==0.7.5
traitlets==4.0.0
uritemplate==0.6
wcwidth==0.1.6
Werkzeug==0.11.2
